+++ binaryen/binaryen/scripts/test/spectest.js
@@ -1,3 +1,0 @@
-export function print(...args) {
-  console.log(...args);
-}