+++ ghc-toolkit/boot-libs/base/GHC/Show.hs
@@ -1,7 +1,7 @@
 {-# LANGUAGE Trustworthy #-}
 {-# LANGUAGE CPP, NoImplicitPrelude, BangPatterns, StandaloneDeriving,
              MagicHash, UnboxedTuples #-}
-{-# OPTIONS_HADDOCK hide #-}
+{-# OPTIONS_HADDOCK not-home #-}
 
 #include "MachDeps.h"
 #if SIZEOF_HSWORD == 4
@@ -53,7 +53,6 @@ import GHC.Base
 import GHC.List ((!!), foldr1, break)
 import GHC.Num
 import GHC.Stack.Types
-import GHC.Types (TypeLitSort (..))
 
 
 -- | The @shows@ functions return a function that prepends the