+++ ghc-toolkit/boot-libs/base/GHC/Unicode.hs
@@ -1,6 +1,6 @@
 {-# LANGUAGE Trustworthy #-}
 {-# LANGUAGE CPP, NoImplicitPrelude, StandaloneDeriving #-}
-{-# OPTIONS_HADDOCK hide #-}
+{-# OPTIONS_HADDOCK not-home #-}
 
 -----------------------------------------------------------------------------
 -- |
@@ -71,7 +71,7 @@ import GHC.Show ( Show )
 -- >>> enumFromTo ModifierLetter SpacingCombiningMark
 -- [ModifierLetter,OtherLetter,NonSpacingMark,SpacingCombiningMark]
 --
--- 'Read' instance:
+-- 'Text.Read.Read' instance:
 --
 -- >>> read "DashPunctuation" :: GeneralCategory
 -- DashPunctuation